Split constants into modules
