+++ morphac/constants/CMakeLists.txt
@@ -12,6 +12,9 @@ set(CONSTANTS_BINARY_DIR ${MORPHAC_BINARY_DIR}/constants)
 set(CONSTANTS_BINDING_DIR ${CONSTANTS_DIR}/binding)
 set(CONSTANTS_BINDING_INCLUDE_DIR ${CONSTANTS_BINDING_DIR}/include)
 set(CONSTANTS_BINDING_SRC_DIR ${CONSTANTS_BINDING_DIR}/src)
+set(CONSTANTS_BINDING_MODULES_DIR
+  ${CONSTANTS_BINDING_DIR}/modules
+)
 set(CONSTANTS_BINDING_BINARY_DIR ${CONSTANTS_BINARY_DIR}/binding)
 set(CONSTANTS_BINDING_INSTALL_DIR
   ${PYTHON_BINDING_INSTALL_DIR}/constants
