+++ morphac/constants/binding/CMakeLists.txt
@@ -6,6 +6,14 @@ if(NOT BUILD_PYTHON_BINDINGS)
   return()
 endif()
 
+
+# Adding subdirectories
+# -------------------------------------------------
+
+add_subdirectory(modules)
+
+
+
 # Src file that creates the python module.
 set(CONSTANTS_MODULE_FILE
