--- morphac/constants/binding/__init__.py
@@ -1,9 +1,0 @@
-from ._binding_constants_python import (
-    # Environment constants.
-    MapConstants,
-    # Mechanics constants.
-    AckermannModelConstants,
-    DiffdriveModelConstants,
-    DubinModelConstants,
-    TricycleModelConstants,
-)
