+++ morphac/constants/binding/modules/environment_constants/__init__.py
@@ -1,0 +1,3 @@
+from morphac.constants._binding_constants_python import (
+    MapConstants,
+)
