+++ morphac/visualization/python/map_visualization.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-from morphac.constants import MapConstants
 from morphac.constants.colors import MapColors
+from morphac.constants.environment_constants import MapConstants
 from morphac.utils.canvas_utils import paint_canvas
